Evaluate drag expression only when drag starts.
This change makes it so the drag expression is evaluated only once when
the dragging start,  instead of $watching it.
This way,  I can use expressions such as  
   drag="[myobj]"
   drag="something + other"
That create an object every time they are evaluated.

This also means a performance win,  since $watching is very expensive,
and completely unnecesary when the user is not doing drag and drop.